refactor: Simplify global-styles file path
The global-styles file has been moved up one directory level, removing an unnecessary sub-directory. This revision improves file organization and path simplicity.
