+++ src/routes/layout.tsx
@@ -2,7 +2,7 @@ import { component$, Slot } from "@builder.io/qwik";
 
 import { Footer } from "~/components/interface/footer";
 import { Header } from "~/components/interface/header";
-import { useGlobalStyles } from "~/hooks/global-styles/global-styles";
+import { useGlobalStyles } from "~/hooks/global-styles";
 
 /** @private */
 export default component$(() => {
